+++ doc/whatsnew.rst
@@ -68,6 +68,8 @@ v0.2.8.dev
 
 - Fix dispersion when generating datasets: :func:`pyriemann.datasets.sample_gaussian_spd`
 
+- Improve base functions to process ndarrays of SPD matrices
+
 v0.2.7 (June 2021)
 ------------------